Шаблонизируем проект. Изменения
